requirements.txt
torch~=1.13.1
torchvision~=0.14.1
torchaudio~=0.13.1
pillow~=9.4.0
transformers~=4.25.1
datasets~=2.8.0
pandas~=1.5.3
tqdm~=4.64.1
scikit-learn~=1.2.0